+++ tgbotapi.behaviour_builder.fsm/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/BehaviourWithFSMStateHandler.kt
@@ -2,6 +2,12 @@ package dev.inmo.tgbotapi.extensions.behaviour_builder
 
 import dev.inmo.micro_utils.fsm.common.*
 
-fun interface BehaviourWithFSMStateHandler<I : O, O : State> {
+fun interface BehaviourWithFSMStateHandler<I : O, O : State> : StatesHandler<I, O> {
     suspend fun BehaviourContextWithFSM<in O>.handleState(state: I): O?
+
+    override suspend fun StatesMachine<in O>.handleState(state: I): O? = if (this is BehaviourContextWithFSM) {
+        handleState(state)
+    } else {
+        null
+    }
 }